RDCC-233: Code committed to fix functional tets case with authrization header
