+++ src/functionalTest/java/uk/gov/hmcts/reform/professionalapi/AuthorizationFunctionalTest.java
@@ -1,35 +1,114 @@
 package uk.gov.hmcts.reform.professionalapi;
 
-import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
-import static com.github.tomakehurst.wiremock.client.WireMock.equalTo;
-import static com.github.tomakehurst.wiremock.client.WireMock.get;
-import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
+import static org.assertj.core.api.Assertions.assertThat;
 
-import com.github.tomakehurst.wiremock.junit.WireMockRule;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import lombok.extern.slf4j.Slf4j;
+import net.serenitybdd.junit.spring.integration.SpringIntegrationSerenityRunner;
+import org.junit.After;
 import org.junit.Before;
-import org.junit.Rule;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.TestPropertySource;
+import uk.gov.hmcts.reform.professionalapi.client.ProfessionalApiClient;
+import uk.gov.hmcts.reform.professionalapi.client.S2sClient;
+import uk.gov.hmcts.reform.professionalapi.config.Oauth2;
+import uk.gov.hmcts.reform.professionalapi.config.TestConfigProperties;
+import uk.gov.hmcts.reform.professionalapi.idam.IdamService;
 
-public class AuthorizationFunctionalTest {
-    @Rule
-    public WireMockRule sidam_service = new WireMockRule(5000);
+
+@RunWith(SpringIntegrationSerenityRunner.class)
+@ContextConfiguration(classes = {TestConfigProperties.class, Oauth2.class})
+@ComponentScan("uk.gov.hmcts.reform.professionalapi")
+@TestPropertySource("classpath:application-functional.yaml")
+@Slf4j
+public abstract class AuthorizationFunctionalTest {
+
+    @Value("${s2s-url}")
+    protected String s2sUrl;
+
+    @Value("${s2s-name}")
+    protected String s2sName;
+
+    @Value("${s2s-secret}")
+    protected String s2sSecret;
+
+    @Value("${targetInstance}")
+    protected String professionalApiUrl;
+
+    @Value("${exui.role.hmcts-admin}")
+    protected String hmctsAdmin;
+
+    @Value("${exui.role.pui-user-manager}")
+    protected String puiUserManager;
+
+    @Value("${exui.role.pui-organisation-manager}")
+    protected String puiOrgManager;
+
+    @Value("${exui.role.pui-finance-manager}")
+    protected String puiFinanceManager;
+
+    @Value("${exui.role.pui-case-manager}")
+    protected String puiCaseManager;
+    
+
+    protected ProfessionalApiClient professionalApiClient;
+
+    protected IdamService idamService;
+
+    @Autowired
+    protected TestConfigProperties configProperties;
+
 
     @Before
-    public void setupIdamStubs() throws Exception {
+    public void setUp() {
+        log.info("Configured S2S secret: " + s2sSecret.substring(0, 2) + "************" + s2sSecret.substring(14));
+        log.info("Configured S2S microservice: " + s2sName);
+        log.info("Configured S2S URL: " + s2sUrl);
 
-        sidam_service.stubFor(get(urlEqualTo("/details"))
-              //  .withHeader("Authorization", equalTo("Bearer pui-case-manager-eyJ0eXAiOiJKV1QiLCJ6aXAiOiJOT05FIiwia2lkIjoiYi9PNk92VnYxK3krV2dySDVVaTlXVGlvTHQwPSIsImFsZyI6IlJTMjU2In0.eyJzdWIiOiJzdXBlci51c2VyQGhtY3RzLm5ldCIsImF1dGhfbGV2ZWwiOjAsImF1ZGl0VHJhY2tpbmdJZCI6IjZiYTdkYTk4LTRjMGYtNDVmNy04ZjFmLWU2N2NlYjllOGI1OCIsImlzcyI6Imh0dHA6Ly9mci1hbTo4MDgwL29wZW5hbS9vYXV0aDIvaG1jdHMiLCJ0b2tlbk5hbWUiOiJhY2Nlc3NfdG9rZW4iLCJ0b2tlbl90eXBlIjoiQmVhcmVyIiwiYXV0aEdyYW50SWQiOiI0NjAzYjVhYS00Y2ZhLTRhNDQtYWQzZC02ZWI0OTI2YjgxNzYiLCJhdWQiOiJteV9yZWZlcmVuY2VfZGF0YV9jbGllbnRfaWQiLCJuYmYiOjE1NTk4OTgxNzMsImdyYW50X3R5cGUiOiJhdXRob3JpemF0aW9uX2NvZGUiLCJzY29wZSI6WyJhY3IiLCJvcGVuaWQiLCJwcm9maWxlIiwicm9sZXMiLCJjcmVhdGUtdXNlciIsImF1dGhvcml0aWVzIl0sImF1dGhfdGltZSI6MTU1OTg5ODEzNTAwMCwicmVhbG0iOiIvaG1jdHMiLCJleHAiOjE1NTk5MjY5NzMsImlhdCI6MTU1OTg5ODE3MywiZXhwaXJlc19pbiI6Mjg4MDAsImp0aSI6IjgxN2ExNjE0LTVjNzAtNGY4YS05OTI3LWVlYjFlYzJmYWU4NiJ9.RLJyLEKldHeVhQEfSXHhfOpsD_b8dEBff7h0P4nZVLVNzVkNoiPdXYJwBTSUrXl4pyYJXEhdBwkInGp3OfWQKhHcp73_uE6ZXD0eIDZRvCn1Nvi9FZRyRMFQWl1l3Dkn2LxLMq8COh1w4lFfd08aj-VdXZa5xFqQefBeiG_xXBxWkJ-nZcW3tTXU0gUzarGY0xMsFTtyRRilpcup0XwVYhs79xytfbq0WklaMJ-DBTD0gux97KiWBrM8t6_5PUfMDBiMvxKfRNtwGD8gN8Vct9JUgVTj9DAIwg0KPPm1rEETRPszYI2wWvD2lpH2AwUtLBlRDANIkN9SdfiHSETvoQ"))
-                .willReturn(aResponse()
-                        .withStatus(200)
-                        .withHeader("Content-Type", "application/json")
-                        .withBody("{" +
-                                "  \"id\": \"ef4fac86-d3e8-47b6-88a7-c7477fb69d3f\","
-                                +  "  \"forename\": \"Super\","
-                                +  "  \"surname\": \"User\","
-                                +  "  \"email\": \"super.user@hmcts.net\","
-                                +  "  \"accountStatus\": \"active\","
-                                +  "  \"roles\": ["
-                                +  "    \"SuperUser\""
-                                +  "  ]"
-                                +  "}")));
+
+        String s2sToken = new S2sClient(s2sUrl, s2sName, s2sSecret).signIntoS2S();
+
+        idamService = new IdamService(configProperties);
+
+        String userToken = idamService.createUserWith("", "SuperUser").getAuthorisationToken();
+
+        professionalApiClient = new ProfessionalApiClient(
+                                                          professionalApiUrl,
+                                                          s2sToken, userToken);
+    }
+
+    @After
+    public void tearDown() {
+    }
+
+    protected String createAndUpdateOrganisationToActive() {
+
+        Map<String, Object> response = professionalApiClient.createOrganisation();
+        String organisationIdentifier = (String) response.get("organisationIdentifier");
+        assertThat(organisationIdentifier).isNotEmpty();
+        professionalApiClient.updateOrganisation(organisationIdentifier,puiCaseManager);
+        return organisationIdentifier;
+
+    }
+
+    protected void validateUsers(Map<String, Object> searchResponse) {
+        assertThat(searchResponse.get("users")).asList().isNotEmpty();
+
+        List<HashMap> professionalUsersResponses = (List<HashMap>) searchResponse.get("users");
+        HashMap professionalUsersResponse = professionalUsersResponses.get(0);
+
+        assertThat(professionalUsersResponse.get("userIdentifier")).isNotNull();
+        assertThat(professionalUsersResponse.get("firstName")).isNotNull();
+        assertThat(professionalUsersResponse.get("lastName")).isNotNull();
+        assertThat(professionalUsersResponse.get("email")).isNotNull();
+        assertThat(professionalUsersResponse.get("status")).isNotNull();
+        assertThat(((List)professionalUsersResponse.get("roles")).size()).isEqualTo(0);
     }
 }
