+++ src/functionalTest/java/uk/gov/hmcts/reform/professionalapi/EndpointSecurityTest.java
@@ -9,7 +9,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
 @ActiveProfiles("functional")
-public class EndpointSecurityTest extends FunctionalTestSuite {
+public class EndpointSecurityTest extends AuthorizationFunctionalTest {
 
     @Test
     public void should_allow_unauthenticated_requests_to_welcome_message_and_return_200_response_code() {
