+++ src/functionalTest/java/uk/gov/hmcts/reform/professionalapi/FindUsersByOrganisationTest.java
@@ -3,6 +3,7 @@ package uk.gov.hmcts.reform.professionalapi;
 import java.util.Map;
 
 import net.serenitybdd.junit.spring.integration.SpringIntegrationSerenityRunner;
+
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.http.HttpStatus;
@@ -11,7 +12,8 @@ import org.springframework.test.context.ActiveProfiles;
 
 @RunWith(SpringIntegrationSerenityRunner.class)
 @ActiveProfiles("functional")
-public class FindUsersByOrganisationTest extends FunctionalTestSuite {
+public class FindUsersByOrganisationTest extends AuthorizationFunctionalTest {
+
 
     @Test
     public void find_users_by_active_organisation_with_showDeleted_False() {
@@ -32,7 +34,7 @@ public class FindUsersByOrganisationTest extends FunctionalTestSuite {
     public void find_users_for_non_active_organisation() {
         Map<String, Object> response = professionalApiClient.createOrganisation();
         String organisationIdentifier = (String) response.get("organisationIdentifier");
-        Map<String, Object> searchResponse = professionalApiClient.searchUsersByOrganisation(organisationIdentifier, puiCaseManager,"False", HttpStatus.NOT_FOUND);
+        Map<String, Object> searchResponse = professionalApiClient.searchUsersByOrganisation(organisationIdentifier, puiCaseManager,"False", HttpStatus.OK);
     }
 
     @Test
