+++ src/functionalTest/java/uk/gov/hmcts/reform/professionalapi/OrganisationCreationsTest.java
@@ -10,7 +10,7 @@ import org.springframework.test.context.ActiveProfiles;
 
 @RunWith(SpringIntegrationSerenityRunner.class)
 @ActiveProfiles("functional")
-public class OrganisationCreationsTest extends FunctionalTestSuite {
+public class OrganisationCreationsTest extends AuthorizationFunctionalTest {
 
 
     @Test
