+++ src/functionalTest/java/uk/gov/hmcts/reform/professionalapi/UserEmailSearchTest.java
@@ -17,7 +17,7 @@ import uk.gov.hmcts.reform.professionalapi.controller.request.OrganisationCreati
 
 @RunWith(SpringIntegrationSerenityRunner.class)
 @ActiveProfiles("functional")
-public class UserEmailSearchTest extends FunctionalTestSuite {
+public class UserEmailSearchTest extends AuthorizationFunctionalTest {
 
     @Test
     public void can_find_a_user_by_their_email_address() {
