+++ src/functionalTest/java/uk/gov/hmcts/reform/professionalapi/idam/IdamApi.java
@@ -5,6 +5,9 @@ import feign.Body;
 import feign.Headers;
 import feign.Param;
 import feign.RequestLine;
+
+import java.util.List;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -12,7 +15,6 @@ import lombok.Getter;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
-import java.util.List;
 
 public interface IdamApi {
 
@@ -27,7 +29,7 @@ public interface IdamApi {
     AuthenticateUserResponse authenticateUser(@Param("authorization") String authorization,
                                               @Param("response_type") String responseType,
                                               @Param("client_id") String clientId,
-                                              @Param("redirect_uri") String redirectUri);
+                                              @Param("redirect_uri") String redirectUri, @Param("scope") String scope);
 
     @RequestLine("POST /oauth2/token")
     @Headers("Content-Type: application/x-www-form-urlencoded")
@@ -44,6 +46,7 @@ public interface IdamApi {
     class CreateUserRequest {
         private final String email;
         private final String forename = "John";
+        private final String id = "my_reference_data_client_id";// UUID.randomUUID().toString();
         private final String surname = "Smith";
         private final UserGroup userGroup;
         private final List<Role> roles;
