+++ src/functionalTest/resources/application-functional.yaml
@@ -1,15 +1,16 @@
 targetInstance: ${TEST_URL:http://localhost:8090}
 s2s-url: ${S2S_URL_FOR_TESTS:http://127.0.0.1:8502}
 s2s-name: rd_professional_api
-# Idam oauth2 settings http://23.101.50.80:5000 :https://idam-api.aat.platform.hmcts.ne
+#Idam oauth2 settings http://23.101.50.80:5000 :https://idam-api.aat.platform.hmcts.net
 idam.api.url: ${IDAM_API_URL:http://23.101.50.80:5000}
+#idam.api.url: ${IDAM_API_URL:https://idam-api.aat.platform.hmcts.net}
 generated.user.email.pattern: ${GENERATED_USER_EMAIL_PATTERN:freg-test-user-%s@feemail.com}
 test.url: ${TEST_URL:http://localhost:8000}
-test.user.password: ${TEST_USER_PASSWORD:password}
-
-oauth2.redirect.uri: ${OAUTH2_REDIRECT_URI:https://moneyclaims.aat.platform.hmcts.net}
-oauth2.client.id: ${OAUTH2_CLIENT_ID:cmc_citizen}
-oauth2.client.secret: ${OAUTH2_CLIENT_SECRET:XXXXXXXX}
+test.user.password: ${TEST_USER_PASSWORD:Hmcts123}
+#https://moneyclaims.aat.platform.hmcts.net : cmc_citizen
+oauth2.redirect.uri: ${OAUTH2_REDIRECT_URI:http://test-sidam/v1/organisation}
+oauth2.client.id: ${OAUTH2_CLIENT_ID:my_reference_data_client_id}
+oauth2.client.secret: ${OAUTH2_CLIENT_SECRET:cmcfinancialremedy}
 # Test key - not used in any deployed vaults or services
 s2s-secret: ${S2S_SECRET:SZDUA3L7N32PE2IS}
 spring.flyway.enabled: false
